backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
dj-database-url==3.0.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
fonttools==4.55.3
googlesearch-python==1.3.0
gunicorn==23.0.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.0
networkx==3.4.2
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
psycopg==3.2.4
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
scipy==1.15.1
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.4.0

# Task Queue & Background Jobs
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
kombu==5.4.2

# Additional utilities for subscriptions
python-crontab==3.2.0

# SMS/Communication
twilio==9.3.7

# Email
django-email-verification==0.3.3

# Utils
python-decouple==3.8
qrcode==8.0